refactor: compartmentalize table script
made this look neater and added a constants file
